requirements.txt
gradio
torch
numpy
pillow
os